fixup! support sending structured attributes in browser console logs
